+++ src/filtering.js
@@ -13,9 +13,13 @@ var engine = {};
 engine.whereMacro = function(parentData, expr) {
   if(parentData !== false && ! parentData) { return []; }
 
-  return util.flatten(parentData.filter((x, i) => {
+  return util.flatten(parentData.map((x, i) => {
     this.$index = i;
-    return expr(x)[0];
+    const condition = expr(x);
+    if (condition instanceof Promise) {
+      return condition.then(c => c[0] ? x : []);
+    }
+    return condition[0] ? x : [];
   }));
 };
 
@@ -42,28 +46,37 @@ engine.selectMacro = function(data, expr) {
   }));
 };
 
-engine.repeatMacro = function(parentData, expr) {
+engine.repeatMacro = function(parentData, expr, res = [], unique = {}) {
   if(parentData !== false && ! parentData) { return []; }
 
-  let res = [];
-  const unique = {};
-  const length = parentData.length;
-  for(let i = 0; i < length; ++i) {
-    let newItems = parentData[i];
-    do {
-      newItems = expr(newItems)
-        .filter(item => {
-          const key = hashObject(item);
-          const isUnique = !unique[key];
-          if (isUnique) {
-            unique[key] = true;
-          }
-          return isUnique;
-        });
-    } while (res.length < res.push.apply(res, newItems));
+  let newItems = [].concat(...parentData.map(i => expr(i)));
+  if (newItems.some(i => i instanceof Promise)) {
+    return Promise.all(newItems).then(items => {
+      items = [].concat(...items);
+      if (items.length) {
+        return engine.repeatMacro(getUniqItems(items, unique, res), expr, res, unique);
+      }
+      return res;
+    });
+  } else if (newItems.length) {
+    return engine.repeatMacro(getUniqItems(newItems, unique, res), expr, res, unique);
+  } else {
+    return res;
   }
-  return res;
 };
+
+function getUniqItems(items, unique, res) {
+  const newItems = items.filter(item => {
+    const key = hashObject(item);
+    const isUnique = !unique[key];
+    if (isUnique) {
+      unique[key] = true;
+    }
+    return isUnique;
+  });
+  res.push.apply(res, newItems);
+  return newItems;
+}
 
 //TODO: behavior on object?
 engine.singleFn = function(x) {